+++ app/admin/facebook_share.rb
@@ -1,13 +1,13 @@
 # Comments
-ActiveAdmin.register TwitterShare do
+ActiveAdmin.register FacebookShare do
   config.clear_sidebar_sections!
   controller.authorize_resource
   config.comments = false
-  menu parent: "Items", priority: 12, label: "Twitter Shares", if: lambda{|tabs_renderer|
-    controller.current_ability.can?(:manage, TwitterShare)
+  menu parent: "Items", priority: 13, label: "Facebook Shares", if: lambda{|tabs_renderer|
+    controller.current_ability.can?(:manage, FacebookShare)
   }
   actions  :index, :destroy
-  index title: "Twitter Shares" do
+  index title: "Facebook Shares" do
     column :item
     column :status
     column "Processed At", :processed_at
@@ -16,7 +16,7 @@ ActiveAdmin.register TwitterShare do
   end
   controller do
     def scoped_collection
-       TwitterShare.includes(:item)
+       FacebookShare.includes(:item)
     end
   end
 end